test: consolidate fragmented cluster integration tests into a single file
